BED-4707 moved sqp delete into new sqp file. Moved related test. Return bad request if a user attempts to unshare a query they aren't shared
