+++ cmd/api/src/api/v2/saved_queries_test.go
@@ -20,7 +20,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/google/uuid"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
@@ -1479,207 +1478,6 @@ func TestResources_DeleteSavedQuery(t *testing.T) {
 
 	handler.ServeHTTP(response, req)
 	require.Equal(t, http.StatusNoContent, response.Code)
-}
-
-func TestResources_UnshareSavedQuery(t *testing.T) {
-
-	userId, err := uuid2.NewV4()
-	require.Nil(t, err)
-	userId2, err := uuid2.NewV4()
-	require.Nil(t, err)
-	userId3, err := uuid2.NewV4()
-	require.Nil(t, err)
-
-	endpoint := "/api/v2/saved-queries/{%s}/permissions"
-	savedQueryId := "1"
-
-	t.Run("user can unshare their owned saved query", func(t *testing.T) {
-		var (
-			mockCtrl  = gomock.NewController(t)
-			mockDB    = mocks.NewMockDatabase(mockCtrl)
-			resources = v2.Resources{DB: mockDB}
-		)
-		defer mockCtrl.Finish()
-
-		mockDB.EXPECT().SavedQueryBelongsToUser(gomock.Any(), userId, int64(1)).Return(true, nil)
-
-		payload := v2.UnshareSavedQueryRequest{
-			UserIds: []uuid2.UUID{userId2, userId3},
-		}
-
-		mockDB.EXPECT().DeleteSavedQueryPermissionsForUsers(gomock.Any(), int64(1), gomock.Any()).Return(nil)
-
-		req, err := http.NewRequestWithContext(createContextWithOwnerId(userId), http.MethodDelete, fmt.Sprintf(endpoint, savedQueryId), must.MarshalJSONReader(payload))
-		require.Nil(t, err)
-
-		req.Header.Set(headers.ContentType.String(), mediatypes.ApplicationJson.String())
-		req = mux.SetURLVars(req, map[string]string{api.URIPathVariableSavedQueryID: savedQueryId})
-
-		response := httptest.NewRecorder()
-		handler := http.HandlerFunc(resources.DeleteSavedQueryPermissions)
-
-		handler.ServeHTTP(response, req)
-		assert.Equal(t, http.StatusNoContent, response.Code)
-	})
-
-	t.Run("user can unshare queries they do not own as an admin", func(t *testing.T) {
-		var (
-			mockCtrl  = gomock.NewController(t)
-			mockDB    = mocks.NewMockDatabase(mockCtrl)
-			resources = v2.Resources{DB: mockDB}
-		)
-		defer mockCtrl.Finish()
-
-		payload := v2.UnshareSavedQueryRequest{
-			UserIds: []uuid2.UUID{userId2, userId3},
-		}
-
-		mockDB.EXPECT().DeleteSavedQueryPermissionsForUsers(gomock.Any(), int64(1), gomock.Any()).Return(nil)
-
-		req, err := http.NewRequestWithContext(createContextWithAdminOwnerId(userId), http.MethodDelete, fmt.Sprintf(endpoint, savedQueryId), must.MarshalJSONReader(payload))
-		require.Nil(t, err)
-
-		req.Header.Set(headers.ContentType.String(), mediatypes.ApplicationJson.String())
-		req = mux.SetURLVars(req, map[string]string{api.URIPathVariableSavedQueryID: savedQueryId})
-
-		response := httptest.NewRecorder()
-		handler := http.HandlerFunc(resources.DeleteSavedQueryPermissions)
-
-		handler.ServeHTTP(response, req)
-		assert.Equal(t, http.StatusNoContent, response.Code)
-	})
-
-	t.Run("user errors unsharing query from themselves", func(t *testing.T) {
-		var (
-			mockCtrl  = gomock.NewController(t)
-			mockDB    = mocks.NewMockDatabase(mockCtrl)
-			resources = v2.Resources{DB: mockDB}
-		)
-		defer mockCtrl.Finish()
-
-		payload := v2.UnshareSavedQueryRequest{
-			Self: true,
-		}
-
-		mockDB.EXPECT().DeleteSavedQueryPermissionsForUser(gomock.Any(), int64(1), userId).Return(fmt.Errorf("an error"))
-
-		req, err := http.NewRequestWithContext(createContextWithAdminOwnerId(userId), http.MethodDelete, fmt.Sprintf(endpoint, savedQueryId), must.MarshalJSONReader(payload))
-		require.Nil(t, err)
-
-		req.Header.Set(headers.ContentType.String(), mediatypes.ApplicationJson.String())
-		req = mux.SetURLVars(req, map[string]string{api.URIPathVariableSavedQueryID: savedQueryId})
-
-		response := httptest.NewRecorder()
-		handler := http.HandlerFunc(resources.DeleteSavedQueryPermissions)
-
-		handler.ServeHTTP(response, req)
-		assert.Equal(t, http.StatusInternalServerError, response.Code)
-	})
-
-	t.Run("user can unshare queries shared to them", func(t *testing.T) {
-		var (
-			mockCtrl  = gomock.NewController(t)
-			mockDB    = mocks.NewMockDatabase(mockCtrl)
-			resources = v2.Resources{DB: mockDB}
-		)
-		defer mockCtrl.Finish()
-
-		payload := v2.UnshareSavedQueryRequest{
-			Self: true,
-		}
-
-		mockDB.EXPECT().DeleteSavedQueryPermissionsForUser(gomock.Any(), int64(1), userId).Return(nil)
-
-		req, err := http.NewRequestWithContext(createContextWithAdminOwnerId(userId), http.MethodDelete, fmt.Sprintf(endpoint, savedQueryId), must.MarshalJSONReader(payload))
-		require.Nil(t, err)
-
-		req.Header.Set(headers.ContentType.String(), mediatypes.ApplicationJson.String())
-		req = mux.SetURLVars(req, map[string]string{api.URIPathVariableSavedQueryID: savedQueryId})
-
-		response := httptest.NewRecorder()
-		handler := http.HandlerFunc(resources.DeleteSavedQueryPermissions)
-
-		handler.ServeHTTP(response, req)
-		assert.Equal(t, http.StatusNoContent, response.Code)
-	})
-
-	t.Run("error user unsharing saved query that does not belong to them", func(t *testing.T) {
-		var (
-			mockCtrl  = gomock.NewController(t)
-			mockDB    = mocks.NewMockDatabase(mockCtrl)
-			resources = v2.Resources{DB: mockDB}
-		)
-		defer mockCtrl.Finish()
-
-		mockDB.EXPECT().SavedQueryBelongsToUser(gomock.Any(), userId, int64(1)).Return(false, nil)
-
-		var userIds []uuid.UUID
-		req, err := http.NewRequestWithContext(createContextWithOwnerId(userId), http.MethodDelete, fmt.Sprintf(endpoint, savedQueryId), must.MarshalJSONReader(userIds))
-		require.Nil(t, err)
-
-		req.Header.Set(headers.ContentType.String(), mediatypes.ApplicationJson.String())
-		req = mux.SetURLVars(req, map[string]string{api.URIPathVariableSavedQueryID: savedQueryId})
-
-		response := httptest.NewRecorder()
-		handler := http.HandlerFunc(resources.DeleteSavedQueryPermissions)
-
-		handler.ServeHTTP(response, req)
-		assert.Equal(t, http.StatusUnauthorized, response.Code)
-	})
-
-	t.Run("error database fails while unsharing to users", func(t *testing.T) {
-		var (
-			mockCtrl  = gomock.NewController(t)
-			mockDB    = mocks.NewMockDatabase(mockCtrl)
-			resources = v2.Resources{DB: mockDB}
-		)
-		defer mockCtrl.Finish()
-
-		payload := v2.UnshareSavedQueryRequest{
-			UserIds: []uuid2.UUID{userId2, userId3},
-		}
-
-		mockDB.EXPECT().DeleteSavedQueryPermissionsForUsers(gomock.Any(), int64(1), gomock.Any()).Return(fmt.Errorf("an error"))
-
-		req, err := http.NewRequestWithContext(createContextWithAdminOwnerId(userId), http.MethodDelete, fmt.Sprintf(endpoint, savedQueryId), must.MarshalJSONReader(payload))
-		require.Nil(t, err)
-
-		req.Header.Set(headers.ContentType.String(), mediatypes.ApplicationJson.String())
-		req = mux.SetURLVars(req, map[string]string{api.URIPathVariableSavedQueryID: savedQueryId})
-
-		response := httptest.NewRecorder()
-		handler := http.HandlerFunc(resources.DeleteSavedQueryPermissions)
-
-		handler.ServeHTTP(response, req)
-		assert.Equal(t, http.StatusInternalServerError, response.Code)
-	})
-
-	t.Run("error database fails while checking saved query ownership", func(t *testing.T) {
-		var (
-			mockCtrl  = gomock.NewController(t)
-			mockDB    = mocks.NewMockDatabase(mockCtrl)
-			resources = v2.Resources{DB: mockDB}
-		)
-		defer mockCtrl.Finish()
-
-		mockDB.EXPECT().SavedQueryBelongsToUser(gomock.Any(), userId, int64(1)).Return(false, fmt.Errorf("an error"))
-
-		payload := v2.UnshareSavedQueryRequest{
-			UserIds: []uuid2.UUID{userId2, userId3},
-		}
-
-		req, err := http.NewRequestWithContext(createContextWithOwnerId(userId), http.MethodDelete, fmt.Sprintf(endpoint, savedQueryId), must.MarshalJSONReader(payload))
-		require.Nil(t, err)
-
-		req.Header.Set(headers.ContentType.String(), mediatypes.ApplicationJson.String())
-		req = mux.SetURLVars(req, map[string]string{api.URIPathVariableSavedQueryID: savedQueryId})
-
-		response := httptest.NewRecorder()
-		handler := http.HandlerFunc(resources.DeleteSavedQueryPermissions)
-
-		handler.ServeHTTP(response, req)
-		assert.Equal(t, http.StatusInternalServerError, response.Code)
-	})
 }
 
 func createContextWithOwnerId(id uuid2.UUID) context.Context {
